src/requirements.txt
ambient_api>=1.5
click >= 8.1
prometheus_client>=0.14
yaml_config_wrapper >=1.0

